Update calcZMixMFE.R
Near the end of the script there is a line of code that calculates zMix by taking the mean of two depths - the depth given by whichMin and the depth given by whichMin+1. These two depths should have been passed to mean() as a vector, within c() - otherwise mean() doesn't calculate the mean of the two. This is now fixed.